Never revoke datastore read permissions
When a private dataset resource table is created in the datastore the
access of the read only datastore user is revoked. From what I can tell
the only reason for this is to stop API users from accessing the data
through the datastore sql search action as this is unauthenticated.
This commit removes the revoke and therefore undoes this protection to
allow other areas of the ckan code to work with private datasets.
Because of this the datastore's SQL search API is also disabled to
protect the private datasets.