Resolve an issue with filter selection
Due to the active setting only existing at the facet level not at the
facet value level, if you selected a facet value to filter by but then
there were other facet values in that facet that could be further
filtered by then you'd end up with some odd looking UI where all of the
facet values were marked as active. This has come about because of the
more accepting backend which is happy with multiple values in all
fields and therefore multiple values for one facet can be used in
combination. To fix this an active setting has been added to the facet
values as well as the facet itself.