Include optional versions in the record URL pages and object links
This is a little messy due to the annoyances of pylons/our slightly
funky implementation due to allowing version to be None but wanting it
to not show up in the URLs if it is None. I may look at this again but
this works for now and all of this will change somewhat with the upgrade
to flask anyway.