Redirect header based rdf requests, don't just serve them
We should be redirecting when we get a request to the object URL of a
record with an Accept header other than html (i.e. an rdf request)
instead of what we're doing at the moment which is just returning the
data. This implements that and cleans it up slightly too.