Only show the fields at the requested version in the slick grid
If a user is viewing an old version of the data and in newer versions of
the data new columns have been added, the user will see these new
columns in the slick grid header row (no data will be shown for them
because the column doesn't exist in the old records). To remove them we
use the hiddenColumns slick option to hide them away. This is a bit of a
hack as really slick grid shouldn't be showing these new columns.