Fix chrome breadcrumb multiline formatting bug
Not really sure why this is a problem, but when the <li> and <a>
tags are formatted nicely with new lines, these new lines are read
by chrome and used resulting in a normal looking breadcrumb until
you get to the resource name which is on the line above the rest
of the breadcrumb. This is presumably an issue on all blink engined
browsers but I haven't checked. From some searching around it seems
that it could be a blink bug or it could be some other bit of css
having an unexpected effect. Either way, this fixes it.