Load up the jinja2 extension before any requests are accepted
The taxonomy extension used to be loaded in the base record view which
meant that if you went directly to the dwc or gbif page, you'd get an
error because the extension wasn't loaded and it was expected to be.
This commit fixes that by loading the extension before any request is
completed.