Add another tab and then update display accordingly
Now we have a specimens tab, a collections tab and an everything tab.
When the specimens tab is selected the title says the specimen count.
When the collections tab is selected the title says the specimen + index
lots + artefacts count. When "everything" is selected the total records
in all resources is shown.

This commit changes the cached collection stats format! You will need
to flush your cache to avoid errors when you first use it.