Fix /specimen citation urls
In Jan 2015 we added citation urls for specimen records under a url prefix
of /specimen/. In Nov/Dec 2015 we changed this to use /object/ instead but
maintained a redirect on the /specimen/ urls to ensure the old ones were
still resolvable. During the ckan upgrade a bug was introduced into this
redirect system, this commit fixes that problem.