Replace the bits of functionality from ckanext-spatial we must have
ckanext-spatial hasn't been updated to work with python3 yet which
means until that happens we can't use it. This is both annoying and
also not the end of the world as we actually use very little of the
functionality it provides.

The bits of functionality we use (afaik) are:

    - datasets can have an extent specified in geojson
    - dataset extents are shown on the package's page (in the sidebar)
    - specimen records with a lat and lon have a map shown on their
      record pages
    - you can perform geo searches using the API to find datasets in
      the provided area

Of these features, only the record page maps are critical. This is
because, currently, we've locked down the ability to add an extent
to a dataset to only admin users and therefore the only dataset with
an extent is the specimen collection. This makes both the ability to
add an extent to a dataset and the ability to query datasets based on
their extent pretty useless currently. Hence, the record page map is
the only useful feature we use right now.

Therefore, this commit implements a record page map in much the same
way as we did before using ckanext-spatial but with some of the fat
trimmed out (e.g. there aren't any config settings available because
only we're going to use this extension so it might as well be hard
coded in).

When, in the future, ckanext-spatial is made python3 compatible we
can evaluate whether we want to bring it back into the fold, however,
it is worth noting that not having this extension does reduce the
complexity of our postgresql install as we don't need the postgis
extension* anymore.

*not having to include the postgis extension is very handy for running
tests as it solves this issue: https://github.com/ckan/ckan/issues/4764
where cleaning the database automatically between tests using the
"clean_db" fixture fails because it tried to delete the spatial_ref_sys
table and can't.