Fix records stats page graph
The stats records page uses the stats extension in ckan core to render the records
graph whereas the contributor and resources pages use ckanext-graph. Both of these
extensions use flot to render the graphs. For the ckanext-graph ones, flot is loaded
along with some additional plugins, these plugins are not present in the stats flot
bundle. We had a stats bundle asset load in the base html which meant we were getting
flot plugin clashes where the ckanext-graph flot with the plugins was being loaded
first and then the stats flot without the plugins was being loaded over the top. This
then broke the ckanext-graph flot plots.

The solution is to remove the stats asset load from the base template and load it in
the template that needs it (records.html), however, this template does need the plugins
used by ckanext-graph but they're not available so we then also need to load those up
which is now done in a bundle in ckanext-nhm called ckanext-nhm/stats-extra.

Yay...