Remove unnecessary reassignment in doi metadata dict
descriptions = metadata_dict.get('descriptions', []) gets a ref to the
value so we don't need to set it again after modifying the list in place.
Additionally, setting it when it didn't exist in the first place results
in the creation setting of 'descriptions' to [] in the metadata_dict
which seems like an unwanted consequence to me.