Alter mam image urls to point to new service
This is the first stage of deployment and is intended to be:

  - easy to test
  - easy to revert

This will not stay like this - the image URLs will be written into
the data importer and therefore be a part of the record. At the point
when that happens we can change the code that handles mam urls to be
code that handles mss image urls.