Convert bbcm.html into a proper template
It was basically a pure html page, didn't inherit anything and therefore
couldn't use the jinja asset tag to load the css and js it needed. For
some reason this worked before but since moving to webassets it doesn't,
so this fixes the problem by making it a proper managed template page
thing.