Add cached specimen by uuid function and use
The helper may call get the record for the same specimen UUID
multiple times so a) split the logic out so that we don't get
the resource too (we don't need it here) and cache it for 5 mins.
5 mins is pretty arbitrary. Given our specimens only can change
once a day this could be much longer but we're really mainly
trying to catch repeated loads of the same specimen on a page
and subsequent reloads.