Refactor vfactor app to use a Vuex store and improve API interfacing
There are a few other quality of life changes in here too but the key
changes are that a Vuex store is now used to track the current query
and the current record. This allows better separation of concerns
between the components and removes the need for events to be emitted
up and state to be shared down the tree.