fix: include base/view-filters in all assets that require it to ensure ckan.views is available
This fixes the old gallery view not being interactive. I'm not exactly sure why but I think it caused some of our code to fail to initialise properly and then the click events were just being missed. Not sure why this suddenly stopped working either? Bit weird but heyho.
It is likely not all of these assets need base/view-filters, I just added it to every asset which included a JS file which references ckan.views to be safe.

Closes: #723