fix: allow json-ld as an rdf format
The ckanext-dcat library allows json-ld as a format but in its CONTENT_TYPES map it doesn't include it. This could be a bug in ckanext-rdf but we can resolve it easily enough, and it allows us to override and add more options into this content types map if we like without polluting upstream.