feat: update nhm plugin to work with vds vNext
BREAKING CHANGE: update nhm plugin to work with vds vNext

Additionally, this removes the record_show action from this plugin and moves it to the vds plugin (where it still can be called as record_show given it aligns well with ckan's naming conventions, but also can be called as vds_data_get to align with vds's naming conventions). There is also a new action, datastore_search, which chains the datastore_search action from vds (or others!) and adds the include_urls parameter.