fix: consider _inclusive fields true by default
These fields are true by default in the backend, but the search UI was treating them as false. As a side effect, this commit also adds the ability to set non-null defaults for all term/comparison types. It also fixes a bug we hadn't noticed where anything that evaluates to false would be set to when editing an existing term (e.g. false or 0).

Also removes a `comparison === 'contains'` line that I probably copied from the text term and forgot to delete.

Closes: #804